requirements.txt
streamlit
langchain
langchain-community
langchain-chroma
langchain-huggingface
langchain-groq
chromadb
sentence-transformers
pypdf
python-dotenv
langchain-text-splitters